requirements.txt
streamlit
google-generativeai
python-dotenv
pathlib
google-api-python-client